requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.11.0
celery>=5.0.5
redis>=3.5.3
psycopg2>=2.8.6
numpy
pandas
